+++ app/src/main/java/zmuzik/slidingpuzzle/model/Tile.java
@@ -1,4 +1,4 @@
-package zmuzik.slidingpuzzle;
+package zmuzik.slidingpuzzle.model;
 
 import android.graphics.Bitmap;
